Fix shit

+++ std/algorithm/ft_lsitoa.c
@@ -25,7 +25,7 @@ char		*ft_lsitoa(long n, char *base)
 	bl = ft_strlen(base);
 	neg = bl == 10 && n < 0;
 	m[0] = neg ? '-' : 0;
-	p = ((n >= 0) ? n : ((~(unsigned long)n) + 1));
+	p = ((n >= 0) ? (unsigned long)n : ((~(unsigned long)n) + 1));
 	l = ft_luitoa(p, base);
 	ft_strcpy(m + neg, l);
 	free(l);
